feat: Introduce Tone of Voice Guide and enhance Diffusion Deep Research components
- Added a comprehensive Tone of Voice Guide for blog posts on addcommitpush.io, outlining core voice attributes and writing style.
- Updated Diffusion Deep Research components, including new prompts and improved report generation features.
- Enhanced blog post content to reflect the new tone guidelines and provide clearer insights into the research process.
- Included new visual elements and structured prompts to improve user engagement and content clarity.
